fix(database): migrations missing types
